Conflicts:
	addons/account/i18n/ja.po
	addons/account/i18n/vi.po
	addons/account_debit_note/i18n/vi.po
	addons/account_fleet/i18n/vi.po
	addons/analytic/i18n/vi.po
	addons/auth_totp/i18n/vi.po
	addons/barcodes/i18n/vi.po
	addons/barcodes_gs1_nomenclature/i18n/vi.po
	addons/base_import/i18n/hr.po
	addons/base_import/i18n/vi.po
	addons/board/i18n/vi.po
	addons/calendar/i18n/vi.po
	addons/crm/i18n/vi.po
	addons/crm_livechat/i18n/vi.po
	addons/event/i18n/ro.po
	addons/event/i18n/vi.po
	addons/event_booth/i18n/vi.po
	addons/fleet/i18n/vi.po
	addons/gamification/i18n/vi.po
	addons/google_calendar/i18n/vi.po
	addons/google_gmail/i18n/vi.po
	addons/google_recaptcha/i18n/vi.po
	addons/hr/i18n/vi.po
	addons/hr_attendance/i18n/vi.po
	addons/hr_contract/i18n/vi.po
	addons/hr_expense/i18n/vi.po
	addons/hr_fleet/i18n/vi.po
	addons/hr_holidays/i18n/vi.po
	addons/hr_holidays_attendance/i18n/vi.po
	addons/hr_recruitment/i18n/vi.po
	addons/hr_recruitment_survey/i18n/vi.po
	addons/hr_skills/i18n/vi.po
	addons/hr_timesheet/i18n/vi.po
	addons/hr_timesheet_attendance/i18n/vi.po
	addons/hr_work_entry/i18n/vi.po
	addons/hr_work_entry_contract/i18n/vi.po
	addons/hr_work_entry_holidays/i18n/vi.po
	addons/im_livechat/i18n/vi.po
	addons/link_tracker/i18n/vi.po
	addons/lunch/i18n/vi.po
	addons/mail/i18n/vi.po
	addons/mail/static/src/core/common/composer.js
	addons/mail/static/tests/tours/mail_composer_test_tour.js
	addons/mail_group/i18n/vi.po
	addons/maintenance/i18n/vi.po
	addons/mass_mailing/i18n/vi.po
	addons/mass_mailing_sms/i18n/vi.po
	addons/mass_mailing_themes/i18n/vi.po
	addons/microsoft_calendar/i18n/vi.po
	addons/microsoft_outlook/i18n/vi.po
	addons/mrp/i18n/vi.po
	addons/mrp_account/i18n/vi.po
	addons/mrp_subcontracting/i18n/vi.po
	addons/mrp_subcontracting_purchase/i18n/vi.po
	addons/onboarding/i18n/vi.po
	addons/partner_autocomplete/i18n/vi.po
	addons/payment_stripe/i18n/vi.po
	addons/payment_worldline/i18n/pt_BR.po
	addons/point_of_sale/i18n/vi.po
	addons/portal/i18n/vi.po
	addons/pos_adyen/i18n/vi.po
	addons/pos_discount/i18n/vi.po
	addons/pos_epson_printer/i18n/vi.po
	addons/pos_hr/i18n/vi.po
	addons/pos_loyalty/i18n/vi.po
	addons/pos_restaurant/i18n/vi.po
	addons/pos_restaurant_adyen/i18n/vi.po
	addons/pos_sale/i18n/vi.po
	addons/pos_six/i18n/vi.po
	addons/pos_stripe/i18n/vi.po
	addons/privacy_lookup/i18n/vi.po
	addons/product/i18n/vi.po
	addons/project/i18n/vi.po
	addons/project_timesheet_holidays/i18n/vi.po
	addons/purchase/i18n/ja.po
	addons/purchase/i18n/vi.po
	addons/purchase_mrp/i18n/vi.po
	addons/purchase_requisition/i18n/vi.po
	addons/purchase_stock/i18n/vi.po
	addons/rating/i18n/vi.po
	addons/repair/i18n/vi.po
	addons/resource/i18n/vi.po
	addons/sale/i18n/vi.po
	addons/sale_crm/i18n/vi.po
	addons/sale_expense/i18n/vi.po
	addons/sale_management/i18n/vi.po
	addons/sale_purchase/i18n/vi.po
	addons/sale_stock/i18n/vi.po
	addons/sale_timesheet/i18n/vi.po
	addons/sales_team/i18n/vi.po
	addons/sms/i18n/vi.po
	addons/spreadsheet/i18n/vi.po
	addons/spreadsheet_dashboard/i18n/vi.po
	addons/spreadsheet_dashboard_im_livechat/i18n/vi.po
	addons/spreadsheet_dashboard_pos_hr/i18n/vi.po
	addons/stock/i18n/vi.po
	addons/stock_account/i18n/vi.po
	addons/stock_delivery/i18n/vi.po
	addons/stock_landed_costs/i18n/vi.po
	addons/stock_picking_batch/i18n/vi.po
	addons/stock_sms/i18n/vi.po
	addons/survey/i18n/vi.po
	addons/web/i18n/az.po
	addons/web/i18n/vi.po
	addons/web/static/src/module_loader.js
	addons/web_editor/i18n/vi.po
	addons/website/i18n/ja.po
	addons/website/i18n/vi.po
	addons/website_blog/i18n/vi.po
	addons/website_event/i18n/vi.po
	addons/website_event_exhibitor/i18n/vi.po
	addons/website_event_track/i18n/vi.po
	addons/website_forum/i18n/vi.po
	addons/website_hr_recruitment/i18n/ru.po
	addons/website_hr_recruitment/i18n/vi.po
	addons/website_livechat/i18n/vi.po
	addons/website_mass_mailing/i18n/vi.po
	addons/website_sale/i18n/pt.po
	addons/website_sale/i18n/vi.po
	addons/website_slides/i18n/vi.po
	addons/website_slides_forum/i18n/vi.po
	odoo/addons/base/i18n/vi.po

+++ addons/point_of_sale/models/pos_session.py
@@ -95,6 +95,12 @@ class PosSession(models.Model):
 
     _sql_constraints = [('uniq_name', 'unique(name)', "The name of this POS Session must be unique!")]
 
+    def write(self, vals):
+        if vals.get('state') == 'closed':
+            for record in self:
+                record.config_id._notify(('CLOSING_SESSION', {'login_number': self.env.context.get('login_number', False)}))
+        return super().write(vals)
+
     @api.model
     def _load_pos_data_relations(self, model, response):
         model_fields = self.env[model]._fields
@@ -131,8 +137,8 @@ class PosSession(models.Model):
     @api.model
     def _load_pos_data_models(self, config_id):
         return ['pos.config', 'pos.order', 'pos.order.line', 'pos.pack.operation.lot', 'pos.payment', 'pos.payment.method', 'pos.printer',
-            'pos.category', 'pos.bill', 'res.company', 'account.tax', 'account.tax.group', 'product.product', 'product.attribute', 'product.attribute.custom.value',
-            'product.template.attribute.line', 'product.template.attribute.value', 'product.combo', 'product.combo.item', 'product.packaging', 'res.users', 'res.partner',
+                        'pos.category', 'pos.bill', 'res.company', 'account.tax', 'account.tax.group', 'product.product', 'product.template', 'product.template.attribute.line', 'product.attribute',
+            'product.attribute.custom.value', 'product.template.attribute.value', 'product.combo', 'product.combo.item', 'product.packaging', 'res.users', 'res.partner',
             'decimal.precision', 'uom.uom', 'uom.category', 'res.country', 'res.country.state', 'res.lang', 'product.pricelist', 'product.pricelist.item', 'product.category',
             'account.cash.rounding', 'account.fiscal.position', 'account.fiscal.position.tax', 'stock.picking.type', 'res.currency', 'pos.note', 'ir.ui.view', 'product.tag', 'ir.module.module']
 
@@ -187,6 +193,10 @@ class PosSession(models.Model):
 
     def delete_opening_control_session(self):
         self.ensure_one()
+        if not self.exists():
+            return {
+                'status': 'success',
+            }
         if self.state != 'opening_control' or len(self.order_ids) > 0:
             raise UserError(_("You can only cancel a session that is in opening control state and has no orders."))
         self.sudo().unlink()
@@ -195,7 +205,6 @@ class PosSession(models.Model):
         }
 
     def get_pos_ui_product_pricelist_item_by_product(self, product_tmpl_ids, product_ids, config_id):
-        pricelist_fields = self.env['product.pricelist']._load_pos_data_fields(config_id)
         pricelist_item_fields = self.env['product.pricelist.item']._load_pos_data_fields(config_id)
 
         pricelist_item_domain = [
@@ -208,12 +217,8 @@ class PosSession(models.Model):
         ]
 
         pricelist_item = self.env['product.pricelist.item'].search(pricelist_item_domain)
-        pricelist = pricelist_item.pricelist_id
 
-        return {
-            'product.pricelist.item': pricelist_item.read(pricelist_item_fields, load=False),
-            'product.pricelist': pricelist.read(pricelist_fields, load=False)
-        }
+        return {'product.pricelist.item': pricelist_item.read(pricelist_item_fields, load=False)}
 
     @api.depends('currency_id', 'company_id.currency_id')
     def _compute_is_in_company_currency(self):
@@ -226,7 +231,8 @@ class PosSession(models.Model):
             cash_payment_method = session.payment_method_ids.filtered('is_cash_count')[:1]
             if cash_payment_method:
                 total_cash_payment = 0.0
-                result = self.env['pos.payment']._read_group([('session_id', '=', session.id), ('payment_method_id', '=', cash_payment_method.id)], aggregates=['amount:sum'])
+                captured_cash_payments_domain = AND([session._get_captured_payments_domain(),[('payment_method_id', '=', cash_payment_method.id)]])
+                result = self.env['pos.payment']._read_group(captured_cash_payments_domain, aggregates=['amount:sum'])
                 total_cash_payment = result[0][0] or 0.0
                 if session.state == 'closed':
                     total_cash = session.cash_real_transaction + total_cash_payment
@@ -241,7 +247,7 @@ class PosSession(models.Model):
 
     @api.depends('order_ids.payment_ids.amount')
     def _compute_total_payments_amount(self):
-        result = self.env['pos.payment']._read_group([('session_id', 'in', self.ids)], ['session_id'], ['amount:sum'])
+        result = self.env['pos.payment']._read_group(self._get_captured_payments_domain(), ['session_id'], ['amount:sum'])
         session_amount_map = {session.id: amount for session, amount in result}
         for session in self:
             session.total_payments_amount = session_amount_map.get(session.id) or 0
@@ -369,8 +375,6 @@ class PosSession(models.Model):
         # we only open sessions that haven't already been opened
         for session in self.filtered(lambda session: session.state == 'opening_control'):
             values = {}
-            if not session.start_at:
-                values['start_at'] = fields.Datetime.now()
             if session.config_id.cash_control and not session.rescue:
                 last_session = self.search([('config_id', '=', session.config_id.id), ('id', '!=', session.id)], limit=1)
                 session.cash_register_balance_start = last_session.cash_register_balance_end_real  # defaults to 0 if lastsession is empty
@@ -571,7 +575,6 @@ class PosSession(models.Model):
             }
 
         self.post_close_register_message()
-        self.config_id._notify(('CLOSING_SESSION', {'login_number': self.env.context.get('login_number', False)}))
         return {'successful': True}
 
     def post_close_register_message(self):
@@ -629,10 +632,10 @@ class PosSession(models.Model):
     def _get_diff_account_move_ref(self, payment_method):
         return _('Closing difference in %(payment_method)s (%(session)s)', payment_method=payment_method.name, session=self.name)
 
-    def _get_diff_vals(self, payment_method_id, diff_amount):
+    def _get_diff_vals(self, payment_method_id, diff_amount, outstanding_account=False):
         payment_method = self.env['pos.payment.method'].browse(payment_method_id)
         diff_compare_to_zero = self.currency_id.compare_amounts(diff_amount, 0)
-        source_account = payment_method.outstanding_account_id
+        source_account = payment_method.outstanding_account_id or outstanding_account
         destination_account = self.env['account.account']
 
         if (diff_compare_to_zero > 0):
@@ -899,7 +902,7 @@ class PosSession(models.Model):
                 base_lines = order.with_context(linked_to_pos=True)._prepare_tax_base_line_values()
                 AccountTax._add_tax_details_in_base_lines(base_lines, order.company_id)
                 AccountTax._round_base_lines_tax_details(base_lines, order.company_id)
-                AccountTax._add_accounting_data_in_base_lines_tax_details(base_lines, order.company_id)
+                AccountTax._add_accounting_data_in_base_lines_tax_details(base_lines, order.company_id, include_caba_tags=True)
                 tax_results = AccountTax._prepare_tax_lines(base_lines, order.company_id)
                 total_amount_currency = 0.0
                 for base_line, to_update in tax_results['base_lines_to_update']:
@@ -1074,14 +1077,14 @@ class PosSession(models.Model):
             vals.append(self._get_combine_receivable_vals(payment_method, amounts['amount'], amounts['amount_converted']))
         for payment, amounts in split_receivables_pay_later.items():
             vals.append(self._get_split_receivable_vals(payment, amounts['amount'], amounts['amount_converted']))
-        MoveLine.create(vals)
+        data['pay_later_move_lines'] = MoveLine.create(vals)
         return data
 
     def _create_combine_account_payment(self, payment_method, amounts, diff_amount):
         outstanding_account = payment_method.outstanding_account_id
         destination_account = self._get_receivable_account(payment_method)
 
-        account_payment = self.env['account.payment'].create({
+        account_payment = self.env['account.payment'].with_context(pos_payment=True).create({
             'amount': abs(amounts['amount']),
             'journal_id': payment_method.journal_id.id,
             'force_outstanding_account_id': outstanding_account.id,
@@ -1092,10 +1095,18 @@ class PosSession(models.Model):
             'company_id': self.company_id.id,
         })
 
+        # In community the outstanding account is computed on the creation of account.payment records
+        accounting_installed = self.env['account.move']._get_invoice_in_payment_state() == 'in_payment'
+        if not account_payment.outstanding_account_id and accounting_installed:
+            account_payment.outstanding_account_id = account_payment._get_outstanding_account(account_payment.payment_type)
+
         if float_compare(amounts['amount'], 0, precision_rounding=self.currency_id.rounding) < 0:
             # revert the accounts because account.payment doesn't accept negative amount.
-            account_payment.outstanding_account_id = account_payment.destination_account_id
-            account_payment.destination_account_id = account_payment.outstanding_account_id
+            account_payment.write({
+                'outstanding_account_id': account_payment.destination_account_id,
+                'destination_account_id': account_payment.outstanding_account_id,
+                'payment_type': 'outbound',
+            })
 
         account_payment.action_post()
 
@@ -1106,7 +1117,7 @@ class PosSession(models.Model):
         return account_payment.move_id.line_ids.filtered(lambda line: line.account_id == self._get_receivable_account(payment_method))
 
     def _apply_diff_on_account_payment_move(self, account_payment, payment_method, diff_amount):
-        diff_vals = self._get_diff_vals(payment_method.id, diff_amount)
+        diff_vals = self._get_diff_vals(payment_method.id, diff_amount, account_payment.outstanding_account_id)
         if not diff_vals:
             return
         source_vals, dest_vals = diff_vals
@@ -1122,6 +1133,9 @@ class PosSession(models.Model):
                     'credit': new_balance_compare_to_zero < 0 and -new_balance or 0.0
                 })
             ]
+        })
+        account_payment.write({
+            'amount': abs(new_balance),
         })
         account_payment.move_id.action_post()
 
@@ -1278,7 +1292,6 @@ class PosSession(models.Model):
         payment_method_to_receivable_lines = data.get('payment_method_to_receivable_lines')
         payment_to_receivable_lines = data.get('payment_to_receivable_lines')
 
-
         all_lines = (
               split_cash_statement_lines
             | combine_cash_statement_lines
@@ -1296,11 +1309,11 @@ class PosSession(models.Model):
         for payment_method, lines in payment_method_to_receivable_lines.items():
             receivable_account = self._get_receivable_account(payment_method)
             if receivable_account.reconcile:
-                lines.filtered(lambda line: not line.reconciled).reconcile()
+                lines.filtered(lambda line: not line.reconciled).with_context(no_cash_basis=True).reconcile()
 
         for payment, lines in payment_to_receivable_lines.items():
             if payment.partner_id.property_account_receivable_id.reconcile:
-                lines.filtered(lambda line: not line.reconciled).reconcile()
+                lines.filtered(lambda line: not line.reconciled).with_context(no_cash_basis=True).reconcile()
 
         # Reconcile invoice payments' receivable lines. But we only do when the account is reconcilable.
         # Though `account_default_pos_receivable_account_id` should be of type receivable, there is currently
@@ -1308,11 +1321,11 @@ class PosSession(models.Model):
         if self.company_id.account_default_pos_receivable_account_id.reconcile:
             for payment_method in combine_inv_payment_receivable_lines:
                 lines = combine_inv_payment_receivable_lines[payment_method] | combine_invoice_receivable_lines.get(payment_method, self.env['account.move.line'])
-                lines.filtered(lambda line: not line.reconciled).reconcile()
+                lines.filtered(lambda line: not line.reconciled).with_context(no_cash_basis=True).reconcile()
 
             for payment in split_inv_payment_receivable_lines:
                 lines = split_inv_payment_receivable_lines[payment] | split_invoice_receivable_lines.get(payment, self.env['account.move.line'])
-                lines.filtered(lambda line: not line.reconciled).reconcile()
+                lines.filtered(lambda line: not line.reconciled).with_context(no_cash_basis=True).reconcile()
 
         # reconcile stock output lines
         pickings = self.picking_ids.filtered(lambda p: not p.pos_order_id)
@@ -1322,7 +1335,7 @@ class PosSession(models.Model):
         for account_id in stock_output_lines:
             ( stock_output_lines[account_id]
             | stock_account_move_lines.filtered(lambda aml: aml.account_id == account_id)
-            ).filtered(lambda aml: not aml.reconciled).reconcile()
+            ).filtered(lambda aml: not aml.reconciled).with_context(no_cash_basis=True).reconcile()
         return data
 
     def _get_rounding_difference_vals(self, amount, amount_converted):
@@ -1384,10 +1397,10 @@ class PosSession(models.Model):
         else:
             product_name = ""
             product_uom = False
-        title = 'Sales' if sign == 1 else 'Refund'
-        name = '%s untaxed' % title
+        title = _('Sales') if sign == 1 else _('Refund')
+        name = _('%s untaxed', title)
         if applied_taxes:
-            name = '%s %s with %s' % (title, product_name, ', '.join([tax.name for tax in applied_taxes]))
+            name = _('%(title)s %(product_name)s with %(taxes)s', title=title, product_name=product_name, taxes=', '.join([tax.name for tax in applied_taxes]))
         partial_vals = {
             'name': name,
             'account_id': account_id,
@@ -1640,9 +1653,12 @@ class PosSession(models.Model):
             'type': 'ir.actions.act_window',
             'res_model': 'pos.payment',
             'view_mode': 'list,form',
-            'domain': [('session_id', '=', self.id)],
+            'domain': self._get_captured_payments_domain(),
             'context': {'search_default_group_by_payment_method': 1}
         }
+    
+    def _get_captured_payments_domain(self):
+        return [('session_id', 'in', self.ids), ('pos_order_id.state', 'in', ['paid', 'invoiced', 'done'])]
 
     def open_frontend_cb(self):
         """Open the pos interface with config_id as an extra argument.
@@ -1657,7 +1673,10 @@ class PosSession(models.Model):
         return self.config_id.open_ui()
 
     def set_opening_control(self, cashbox_value: int, notes: str):
+        if self.state != 'opening_control':
+            return
         self.state = 'opened'
+        self.start_at = fields.Datetime.now()
         if not self.rescue:
             self.name = self.env['ir.sequence'].with_context(company_id=self.config_id.company_id.id).next_by_code('pos.session')
 
@@ -1899,5 +1918,15 @@ class ProcurementGroup(models.Model):
     def _run_scheduler_tasks(self, use_new_cursor=False, company_id=False):
         super(ProcurementGroup, self)._run_scheduler_tasks(use_new_cursor=use_new_cursor, company_id=company_id)
         self.env['pos.session']._alert_old_session()
+        if 'scheduler_task_done' in self._context:
+            task_done = self._context.get('scheduler_task_done', {'task_done': 0})['task_done'] + 1
+            self._context['scheduler_task_done']['task_done'] = task_done
+        else:
+            task_done = self._get_scheduler_tasks_to_do()
         if use_new_cursor:
+            self.env['ir.cron']._notify_progress(done=task_done, remaining=self._get_scheduler_tasks_to_do() - task_done)
             self.env.cr.commit()
+
+    @api.model
+    def _get_scheduler_tasks_to_do(self):
+        return super()._get_scheduler_tasks_to_do() + 1
